refactor: Export ContextDocumentLoader from utility file
